// samples/swr-with-effect/src/gen/endpoints/pets/pets.effect.ts
export const CreatePetsBodyItem = S.Struct({
  name: S.String,
  tag: S.String,
});